+++ examples/myfsm.py
@@ -59,7 +59,7 @@ class myfsm(fsmBase):
             self.gotoState("uno")    
  
     def common_eval(self):
-        if self.tmrExp("timer"):
+        if self.tmrExpired("timer"):
             self.logI("Timer expired")
             self.gotoState("uno")